only run state from bot if in config

+++ bot.py
@@ -4,14 +4,15 @@ from socketIO_client import SocketIO, BaseNamespace
 import threading
 import sys
 import yaml
-import subprocess as sp
-import os
-
-state = sp.Popen(['python','state.py'])
 
 conf_file = sys.argv[1] if len(sys.argv) > 1 else "config.yml"
 with open(conf_file) as f:
     config = yaml.load(f)
+
+if config["state"]["run_from_bot"]:
+    import subprocess as sp
+    import os
+    state = sp.Popen(['python', 'state.py'])
 
 admins = ["fwilson", "sdrodge", "okulkarni"]
 denied = {}
@@ -70,9 +71,12 @@ class AvalonBot(IRCBot):
                 args = message.split()
                 denied[args[1]] = []
                 bot.send(channel, "ok!")
-            elif message == '!restart':
-                sp.Popen.terminate(state)
-                os.execl(sys.executable, sys.executable, *sys.argv)
+            elif message == '!restart' and nickname in admins:
+                if config["state"]["run_from_bot"]:
+                    sp.Popen.terminate(state)
+                    os.execl(sys.executable, sys.executable, *sys.argv)
+                else:
+                    return
         else:
             message = message.lower()
             if message == "yes":
